# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Simulação de Crédito",
        "description": "API para simulação de crédito com diferentes instituições e convênios",
        "contact": {
            "name": "Suporte API",
            "email": "contato@exemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/convenios": {
            "get": {
                "tags": [
                    "Convênios"
                ],
                "summary": "Lista todos os convênios disponíveis",
                "description": "Retorna a lista completa de convênios que podem ser utilizados para simulação de crédito",
                "operationId": "getConvenios",
                "responses": {
                    "200": {
                        "description": "Lista de convênios retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Convenio"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/v1/convenios/{id}": {
            "get": {
                "tags": [
                    "Convênios"
                ],
                "summary": "Busca um convênio pelo ID",
                "description": "Retorna os dados de um convênio específico identificado pelo ID",
                "operationId": "getConvenioById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do convênio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Convênio encontrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Convenio"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Convênio não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Convênio não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/v1/instituicoes": {
            "get": {
                "tags": [
                    "Instituições"
                ],
                "summary": "Lista todas as instituições financeiras disponíveis",
                "description": "Retorna a lista completa de instituições financeiras que podem ser utilizadas para simulação de crédito",
                "operationId": "getInstituicoes",
                "responses": {
                    "200": {
                        "description": "Lista de instituições retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Instituicao"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/v1/instituicoes/{id}": {
            "get": {
                "tags": [
                    "Instituições"
                ],
                "summary": "Busca uma instituição financeira pelo ID",
                "description": "Retorna os dados de uma instituição financeira específica identificada pelo ID",
                "operationId": "getInstituicaoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instituição encontrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Instituicao"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Instituição não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Instituição não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/v1/simulacao-credito": {
            "post": {
                "tags": [
                    "Simulação de Crédito"
                ],
                "summary": "Realiza uma simulação de crédito",
                "description": "Recebe os parâmetros para simulação e retorna as opções de crédito disponíveis",
                "operationId": "simularCredito",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valor_emprestimo"
                                ],
                                "properties": {
                                    "valor_emprestimo": {
                                        "description": "Valor do empréstimo desejado",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "instituicoes": {
                                        "description": "Lista de instituições para filtrar (opcional)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "banco1",
                                            "banco2"
                                        ]
                                    },
                                    "convenios": {
                                        "description": "Lista de convênios para filtrar (opcional)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "conv1",
                                            "conv2"
                                        ]
                                    },
                                    "parcela": {
                                        "description": "Número de parcelas desejado (opcional)",
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Simulação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SimulacaoCredito"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro na validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos"
                                        },
                                        "errors": {
                                            "properties": {
                                                "valor_emprestimo": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O valor do empréstimo é obrigatório"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Convenio": {
                "title": "Convênio",
                "description": "Modelo de convênio para simulação de crédito",
                "properties": {
                    "id": {
                        "description": "Identificador único do convênio",
                        "type": "string",
                        "example": "INSS"
                    },
                    "valor": {
                        "description": "Nome do convênio",
                        "type": "string",
                        "example": "Instituto Nacional do Seguro Social"
                    }
                },
                "type": "object"
            },
            "Instituicao": {
                "title": "Instituição",
                "description": "Modelo de instituição financeira para simulação de crédito",
                "properties": {
                    "id": {
                        "description": "Identificador único da instituição",
                        "type": "string",
                        "example": "BC001"
                    },
                    "valor": {
                        "description": "Nome da instituição financeira",
                        "type": "string",
                        "example": "Banco X"
                    }
                },
                "type": "object"
            },
            "SimulacaoCredito": {
                "title": "Simulação de Crédito",
                "description": "Resultado de uma simulação de crédito",
                "properties": {
                    "instituicao": {
                        "description": "Nome da instituição financeira",
                        "type": "string",
                        "example": "Banco X"
                    },
                    "convenio": {
                        "description": "Nome do convênio",
                        "type": "string",
                        "example": "INSS"
                    },
                    "parcelas": {
                        "description": "Número de parcelas da simulação",
                        "type": "integer",
                        "example": 12
                    },
                    "taxaJuros": {
                        "description": "Taxa de juros mensal aplicada",
                        "type": "number",
                        "format": "float",
                        "example": 1.99
                    },
                    "valorParcela": {
                        "description": "Valor de cada parcela",
                        "type": "number",
                        "format": "float",
                        "example": 458.33
                    },
                    "valorTotal": {
                        "description": "Valor total a ser pago",
                        "type": "number",
                        "format": "float",
                        "example": 5500
                    },
                    "valorEmprestimo": {
                        "description": "Valor do empréstimo solicitado",
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Convênios",
            "description": "Convênios"
        },
        {
            "name": "Instituições",
            "description": "Instituições"
        },
        {
            "name": "Simulação de Crédito",
            "description": "Simulação de Crédito"
        }
    ]
}